Add landing page, share feature, and layout redesign
