+++ web/messages/en.json
@@ -32,4 +32,5 @@
   "cardView": "Card View",
   "tableView": "Table View",
   "visualAnalysis": "Visual Analysis"
+  ,"share": "Share"
 }
