+++ web/messages/ja.json
@@ -32,4 +32,5 @@
   "cardView": "カード表示",
   "tableView": "テーブル表示",
   "visualAnalysis": "ビジュアル分析"
+  ,"share": "共有"
 }
