+++ web/pages/index.tsx
@@ -1,133 +1,58 @@
-import { useState } from 'react';
-import { useRouter } from 'next/router';
-import { useTranslations } from 'next-intl';
-import LanguageSwitcher from '../components/LanguageSwitcher';
-import CandidateInputs from '../components/CandidateInputs';
-import CriteriaInputs, { Criterion } from '../components/CriteriaInputs';
-import Spinner from '../components/Spinner';
+import Link from 'next/link';
 
 export default function Home() {
-  const t = useTranslations();
-  const router = useRouter();
-  const [candidates, setCandidates] = useState<string[]>(['', '']);
-  const [criteria, setCriteria] = useState<Criterion[]>([{ name: '', weight: 3 }]);
-  const [error, setError] = useState('');
-  const [loading, setLoading] = useState(false);
-
-  const insertSample = () => {
-    setCandidates(['ラーメン屋A', 'ラーメン屋B', 'ラーメン屋C']);
-    setCriteria([
-      { name: '味', weight: 5 },
-      { name: '値段', weight: 4 },
-      { name: 'アクセス', weight: 3 },
-    ]);
-  };
-
-  const apiUrl = process.env.NEXT_PUBLIC_API_URL || 'http://localhost:8000';
-
-  const handleSubmit = async () => {
-    if (candidates.some((c) => !c.trim()) || criteria.some((c) => !c.name.trim())) {
-      setError(t('fillError'));
-      return;
-    }
-    setError('');
-    setLoading(true);
-    const prompt = [
-      `Rank the following items: ${candidates.join(', ')}.`,
-      `Criteria with weights: ${criteria
-        .map((c) => `${c.name} (${c.weight})`)
-        .join(', ')}.`,
-      "Return all candidates as JSON {\"rankings\": [{\"name\":\"\",\"score\":0,\"rank\":0,\"reasons\":{}}]}.",
-      "Reasons must be in Japanese."
-    ].join(' ');
-    console.log('prompt', prompt);
-    try {
-      const res = await fetch(`${apiUrl}/rank`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ prompt, count: candidates.length })
-      });
-      console.log('rank api status', res.status);
-      if (!res.ok) {
-        throw new Error('status ' + res.status);
-      }
-      const data = await res.json();
-      console.log('ranking response', data);
-      let resultArray: any = [];
-      if (Array.isArray(data)) {
-        if (data.length === 1 && data[0]?.rankings) {
-          resultArray = data[0].rankings;
-        } else {
-          resultArray = data;
-        }
-      } else if (Array.isArray(data.results)) {
-        resultArray = data.results;
-      } else if (Array.isArray(data.rankings)) {
-        resultArray = data.rankings;
-      } else {
-        resultArray = [data.results ?? data.rankings ?? data];
-      }
-      if (!resultArray || resultArray.length === 0) {
-        setError(t('noResults'));
-        return;
-      }
-      if (
-        !Array.isArray(resultArray) ||
-        resultArray.some(
-          (r) =>
-            !r || typeof r.name !== 'string' ||
-            r.score === undefined ||
-            r.rank === undefined
-        )
-      ) {
-        setError(t('formatError'));
-        return;
-      }
-      if (typeof window !== 'undefined') {
-        localStorage.setItem('rankingData', JSON.stringify(resultArray));
-      }
-      router.push('/results');
-    } catch (e) {
-      console.error(e);
-      setError(t('fetchError'));
-    } finally {
-      setLoading(false);
-    }
-  };
-
   return (
-    <div className="max-w-[1280px] mx-auto px-4">
-      <div className="max-w-xl mx-auto mt-10 space-y-6">
-        <LanguageSwitcher />
-        <h1 className="text-3xl font-bold text-center">{t('generate')}</h1>
-        <p className="text-center text-sm text-gray-600">{t('instruction')}</p>
-      <div className="text-right">
-        <button
-          type="button"
-          onClick={insertSample}
-          className="text-xs text-blue-600 hover:underline"
-        >
-          {t('sample')}
-        </button>
-      </div>
-      <section className="bg-white p-4 rounded-lg shadow space-y-2">
-        <h2 className="text-xl font-semibold mb-2">{t('candidates')}</h2>
-        <CandidateInputs candidates={candidates} setCandidates={setCandidates} />
+    <div className="max-w-[1100px] mx-auto px-4 text-center space-y-20">
+      <section className="space-y-6 pt-20">
+        <h1 className="text-5xl font-bold">あらゆる『迷い』に、明確な答えを。</h1>
+        <p className="text-lg">複数の選択肢を、あなただけの評価基準で簡単にランキング。直感的なビジュアル分析で、最適な選択をサポートします。</p>
+        <Link href="/create" className="inline-block px-6 py-3 bg-primary text-white rounded-lg hover:bg-primary-dark text-xl">無料でランキングを作成する</Link>
       </section>
-      <section className="bg-white p-4 rounded-lg shadow space-y-2">
-        <h2 className="text-xl font-semibold mb-2">{t('criteria')}</h2>
-        <CriteriaInputs criteria={criteria} setCriteria={setCriteria} />
+      <section className="space-y-8">
+        <h2 className="text-3xl font-bold">たった3ステップで、意思決定をサポート</h2>
+        <div className="grid md:grid-cols-3 gap-6">
+          <div className="space-y-2">
+            <div className="text-6xl">1️⃣</div>
+            <h3 className="font-semibold">入力する</h3>
+            <p>比較したい候補と、あなただけの評価基準を入力します。</p>
+          </div>
+          <div className="space-y-2">
+            <div className="text-6xl">2️⃣</div>
+            <h3 className="font-semibold">評価する</h3>
+            <p>各評価基準の重要度（重み）をスライダーで直感的に設定します。</p>
+          </div>
+          <div className="space-y-2">
+            <div className="text-6xl">3️⃣</div>
+            <h3 className="font-semibold">可視化する</h3>
+            <p>AIが瞬時にスコアを計算し、順位とレーダーチャートで結果をわかりやすく可視化します。</p>
+          </div>
+        </div>
       </section>
-      {error && <p className="text-red-600 text-sm">{error}</p>}
-      <button
-        onClick={handleSubmit}
-        disabled={loading}
-        className="w-full py-2 bg-blue-600 text-white rounded-md shadow hover:bg-blue-700 transition disabled:opacity-50 flex items-center justify-center gap-2"
-      >
-        {loading && <Spinner />}
-        {loading ? t('generating') : t('generate')}
-      </button>
-      </div>
+      <section className="space-y-8">
+        <h2 className="text-3xl font-bold">あなたの意思決定を加速させる機能の数々</h2>
+        <div className="grid md:grid-cols-2 gap-6">
+          <div className="p-4 bg-white rounded shadow">
+            <h3 className="font-semibold mb-2">多角的なビジュアル分析</h3>
+            <p>レーダーチャートで強みと弱みが一目瞭然。</p>
+          </div>
+          <div className="p-4 bg-white rounded shadow">
+            <h3 className="font-semibold mb-2">AIによるサマリー</h3>
+            <p>ランキング結果の要点を自動で文章化。</p>
+          </div>
+          <div className="p-4 bg-white rounded shadow">
+            <h3 className="font-semibold mb-2">保存＆共有</h3>
+            <p>作成したランキングを保存し、URLで共有可能。</p>
+          </div>
+          <div className="p-4 bg-white rounded shadow">
+            <h3 className="font-semibold mb-2">柔軟なカスタマイズ</h3>
+            <p>評価基準や重みを自由に追加・編集。</p>
+          </div>
+        </div>
+      </section>
+      <section className="space-y-6 pb-20">
+        <h2 className="text-3xl font-bold">さあ、あなたの最初のランキングを作りましょう</h2>
+        <Link href="/create" className="inline-block px-6 py-3 bg-primary text-white rounded-lg hover:bg-primary-dark text-xl">無料でランキングを作成する</Link>
+      </section>
     </div>
   );
 }
